Backend/requirements.txt
fastapi
uvicorn
requests
finnhub-python